Add --http-header option to spectool
This patch adds a new misc option --http-header. Which can be used multiple
times to create a list of custom headers to send with the HTTP request
used to download files (patch or sources).

The form is "header:value" or "header: value".

I have found the need for this when needing to send in a gitlab token
along with spectool to download sources:

  spectool -g ../my.spec  --http-header "PRIVATE-TOKEN: <token>"